requirements.txt
certifi==2025.8.3
charset-normalizer==3.4.3
contourpy==1.3.2
cycler==0.12.1
filterpy==1.4.5
fonttools==4.60.0
idna==3.10
kiwisolver==1.4.9
matplotlib==3.10.6
numpy==2.2.6
packaging==25.0
pillow==11.3.0
pyparsing==3.2.5
pyserial==3.5
python-dateutil==2.9.0.post0
requests==2.32.5
scipy==1.15.3
six==1.17.0
urllib3==2.5.0
